+++ static_report/src/components/shared/Tables/TableList/CRTableList/CRTableListColumnsSummary.tsx
